+++ apps/cms-v2/src/modules/bulk-sync/bulk-sync-processor.ts
@@ -96,8 +96,6 @@ export class DatocmsBulkSyncProcessor implements BulkSyncProcessor {
             },
           })
           .then((r) => {
-            console.log(r);
-
             if (hooks.onUploadSuccess) {
               hooks.onUploadSuccess({ variantId: variant.id });
             }
@@ -116,7 +114,7 @@ export const BulkSyncProcessorFactory = {
   create(config: AnyProviderConfigSchemaType): BulkSyncProcessor {
     switch (config.type) {
       case "contentful":
-        return new ContentfulBulkSyncProcessor(config as ContentfulProviderConfigType); // todo should be inferred from switch
+        return new ContentfulBulkSyncProcessor(config);
       case "datocms":
         return new DatocmsBulkSyncProcessor(config);
       default:
