+++ apps/cms-v2/src/modules/contentful/contentful-client.ts
@@ -79,8 +79,6 @@ export class ContentfulClient {
 
     const entry = await env.getEntry(opts.variant.id);
 
-    console.log(entry);
-
     return await entry.delete();
   }
