+++ apps/cms-v2/src/modules/datocms/datocms-config-form.tsx
@@ -225,7 +225,6 @@ const AddFormVariant = () => {
           type: "datocms",
         });
       }}
-      onDelete={console.log}
       defaultValues={{
         authToken: "",
         configName: "",
