+++ apps/cms-v2/src/modules/trpc/trpc-context.ts
@@ -1,7 +1,7 @@
 import * as trpcNext from "@trpc/server/adapters/next";
 import { SALEOR_AUTHORIZATION_BEARER_HEADER, SALEOR_API_URL_HEADER } from "@saleor/app-sdk/const";
 import { inferAsyncReturnType } from "@trpc/server";
-import { getBaseUrl } from "@/get-base-url";
+import { getBaseUrl } from "@/modules/shared/get-base-url";
 
 export const createTrpcContext = async ({ res, req }: trpcNext.CreateNextContextOptions) => {
   const baseUrl = getBaseUrl(req.headers);
